[X] Move Form to directory apart. [X] Move the filters to the book because I want to add some features to the user interface in the future.
