+++ app/types/types.ts
@@ -1,6 +1,7 @@
 import { LucideProps } from "lucide-react";
 import { ComponentType } from "react";
 import { Timestamp } from "firebase/firestore";
+import { ReactNode } from "react";
 
 export type FilterBooks = "to read" | "reading" | "finished" | "all";
 
@@ -43,3 +44,10 @@ export interface IFormTypes {
   status: FilterBooks;
   pagesRead: number;
 }
+
+export type FilterCardProps = {
+  children: ReactNode;
+  selectFilter: () => void;
+  key: FilterBooks;
+  // filter: FilterBooks
+};
